Reading variables from json should only be attempted if the file exists.
This should stop the warning about default values from showing up for
worker pods, and now that we've added the .bkp file in the webserver
the expectation is that at least this one will exist.

GitOrigin-RevId: 1e91b870e992fb733e51f4826399d6af61446801

+++ airflow/www/app.py
@@ -46,21 +46,23 @@ from airflow.utils.net import get_hostname
 
 csrf = CSRFProtect()
 
-try:
-    with open('/home/airflow/gcs/env_var.json', 'r') as env_var_json:
-        env_var_content = json.load(env_var_json)
-        os.environ.update(env_var_content)
-    with open('/home/airflow/gcs/env_var.json.bkp', 'w') as env_var_backup:
-        json.dump(env_var_content, env_var_backup)
-except:
+if os.path.isfile('/home/airflow/gcs/env_var.json') \
+  or os.path.isfile('/home/airflow/gcs/env_var.json.bkp'):
     try:
-        with open('/home/airflow/gcs/env_var.json.bkp', 'r') as env_var_json:
+        with open('/home/airflow/gcs/env_var.json', 'r') as env_var_json:
             env_var_content = json.load(env_var_json)
             os.environ.update(env_var_content)
-        logging.info('Composer Environment Variables were loaded from backup.')
+        with open('/home/airflow/gcs/env_var.json.bkp', 'w') as env_var_backup:
+            json.dump(env_var_content, env_var_backup)
     except:
-        logging.warning('Using default Composer Environment Variables. Overrides '
-                        'have not been applied.')
+        try:
+            with open('/home/airflow/gcs/env_var.json.bkp', 'r') as env_var_json:
+                env_var_content = json.load(env_var_json)
+                os.environ.update(env_var_content)
+            logging.info('Composer Environment Variables were loaded from backup.')
+        except:
+            logging.warning('Using default Composer Environment Variables. Overrides '
+                            'have not been applied.')
 if not conf.getboolean('core', 'unit_test_mode'):
     airflow.plugins_manager = six.moves.reload_module(airflow.plugins_manager)
     airflow.configuration = six.moves.reload_module(airflow.configuration)
